IM PROJECT/requirements.txt
Flask>=3.0
PyYAML>=6.0
psycopg2-binary>=2.9
google-generativeai>=0.7.0

